requirements.txt
Django==2.2.10
simplejson==3.11.1
sqlparse==0.2.2
django-filter==1.0.4
django-statici18n==1.7.0
django-timezone-field==2.1
django-constance[database]==2.3.1
django-logentry-admin==1.0.4
django-admin-sortable2==0.6.21
django-appconf==1.0.2
argon2-cffi==18.3.0
djangorestframework==3.11.0
Markdown==2.6.11
python-memcached==1.59
requests==2.21.0

ipython==6.5.0
lorem==0.1.1
names==0.3.0
transliterate==1.10.2
sha3==0.2.1
Faker==4.0.0
